Move some helpers to abstract-ops
