+++ src/lib/helpers.ts
@@ -16,16 +16,6 @@ export function createDataProperty(o: object, p: string | symbol, v: any) {
   Object.defineProperty(o, p, { value: v, writable: true, enumerable: true, configurable: true });
 }
 
-export function createArrayFromList<T extends any[]>(elements: T): T {
-  // We use arrays to represent lists, so this is basically a no-op.
-  // Do a slice though just in case we happen to depend on the unique-ness.
-  return elements.slice() as T;
-}
-
-export function ArrayBufferCopy(dest: ArrayBuffer, destOffset: number, src: ArrayBuffer, srcOffset: number, n: number) {
-  new Uint8Array(dest).set(new Uint8Array(src, srcOffset, n), destOffset);
-}
-
 export function IsFiniteNonNegativeNumber(v: number): boolean {
   if (IsNonNegativeNumber(v) === false) {
     return false;
@@ -132,16 +122,6 @@ export function PromiseCall<T, A extends any[], R>(F: (this: T, ...args: A) => R
   } catch (value) {
     return promiseRejectedWith(value);
   }
-}
-
-// Not implemented correctly
-export function TransferArrayBuffer<T extends ArrayBufferLike>(O: T): T {
-  return O;
-}
-
-// Not implemented correctly
-export function IsDetachedBuffer(O: ArrayBufferLike): boolean { // eslint-disable-line @typescript-eslint/no-unused-vars
-  return false;
 }
 
 export function ValidateAndNormalizeHighWaterMark(highWaterMark: number): number {
